+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.5
+### is.codion.common.core
+- DefaultProxyBuilder.MethodKey now caches the hashCode.
 
 ## 0.18.4
 ### is.codion